Use seed in random generator

+++ DominantColor/Shared/DominantColors.swift
@@ -88,7 +88,7 @@ public enum GroupingAccuracy {
 public struct DefaultParameterValues {
     public static var maxSampledPixels: Int = 1000
     public static var accuracy: GroupingAccuracy = .medium
-    public static var seed: UInt32 = 3571
+    public static var seed: UInt64 = 3571
     public static var memoizeConversions: Bool = false
 }
 
@@ -118,7 +118,7 @@ public func dominantColorsInImage(
         _ image: CGImage,
         maxSampledPixels: Int = DefaultParameterValues.maxSampledPixels,
         accuracy: GroupingAccuracy = DefaultParameterValues.accuracy,
-        seed: UInt32 = DefaultParameterValues.seed,
+        seed: UInt64 = DefaultParameterValues.seed,
         memoizeConversions: Bool = DefaultParameterValues.memoizeConversions
     ) -> [CGColor] {
